+++ lib/content-services/src/lib/viewer/components/alfresco-viewer.component.html
@@ -183,7 +183,6 @@
                 [viewerType]="viewerType"
                 [fileName]="fileName"
                 [isLoading]="isLoading"
-                [allowGoBack]="allowGoBack"
                 [urlFile]="urlFileContent"
                 [tracks]="tracks"
                 [readOnly]="readOnly">
